just a snapshot
	modified:   src/assign/mod.rs
	modified:   src/assign/propagate.rs
	modified:   src/assign/stack.rs
	modified:   src/cdb/cid.rs
	modified:   src/cdb/db.rs
	modified:   src/cdb/mod.rs
	modified:   src/config.rs
	modified:   src/processor/eliminate.rs
	modified:   src/processor/heap.rs
	modified:   src/solver/conflict.rs
	modified:   src/solver/mod.rs
	modified:   src/solver/restart.rs
	modified:   src/solver/search.rs
	modified:   src/solver/vivify.rs
	modified:   src/state.rs
	modified:   src/types.rs

+++ src/assign/mod.rs
@@ -129,7 +129,7 @@ pub struct AssignStack {
     use_rephase: bool,
     best_assign: bool,
     build_best_at: usize,
-    num_best_assign: f64,
+    num_best_assign: usize,
 
     //
     //## Statistics
